Use swift Result
